upgrade: Rework entirely and add Events
* Split out some core functions
* Ensure we drain queues *before* we migrate
* Add Events
